Edição em massa Adição de comentários Renomeio e criação de classes Otimização do código Reposicionamento de itens Barra de progresso na Tela de Login
