+++ nbproject/project.properties
@@ -68,7 +68,7 @@ jnlp.signed=false
 jnlp.signing=
 jnlp.signing.alias=
 jnlp.signing.keystore=
-main.class=com.trabalhopac.healthyliving.TelaLogin
+main.class=com.trabalhopac.healthyliving.Main
 # Substitui\u00e7\u00e3o opcional do atributo de manifesto da Base de c\u00f3digo; use para impedir que RIAs sejam adaptados
 manifest.custom.codebase=
 # Substitui\u00e7\u00e3o opcional do atributo de manifesto de Permiss\u00f5es (valores suportados: sandbox, all-permissions)
