+++ src/com/trabalhopac/healthyliving/ControleDeCalorias.java
@@ -242,7 +242,7 @@ public class ControleDeCalorias extends javax.swing.JFrame {
 
     private void btnVoltarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnVoltarActionPerformed
         // TODO add your handling code here:
-        new Main().setVisible(true);
+        new MainForm().setVisible(true);
         dispose();
     }//GEN-LAST:event_btnVoltarActionPerformed
